raw string delimiter changed to R json(...)json so the (min 60) inside doesn't prematurely terminate it; cron.cpp: (unsigned) cast so %u matches the argument type

+++ esp32/main/cron.cpp
@@ -29,7 +29,7 @@ esp_err_t CronScheduler::init(MemoryStore* memory) {
 
 std::string CronScheduler::next_id() {
     char buf[16];
-    snprintf(buf, sizeof(buf), "cron_%u", ++id_counter_);
+    snprintf(buf, sizeof(buf), "cron_%u", (unsigned)++id_counter_);
     return std::string(buf);
 }
